IGNITE-3883: Minor fix.

+++ modules/platforms/cpp/odbc/src/app/application_data_buffer.cpp
@@ -1219,7 +1219,7 @@ namespace ignite
                             break;
 
                         std::string str = utility::SqlStringToString(
-                            reinterpret_cast<const unsigned char*>(GetData()), paramLen);
+                            reinterpret_cast<const unsigned char*>(GetData()), static_cast<int32_t>(paramLen));
 
                         sscanf(str.c_str(), "%d-%d-%d %d:%d:%d", &tmTime.tm_year, &tmTime.tm_mon,
                             &tmTime.tm_mday, &tmTime.tm_hour, &tmTime.tm_min, &tmTime.tm_sec);
@@ -1282,7 +1282,7 @@ namespace ignite
                             break;
 
                         std::string str = utility::SqlStringToString(
-                            reinterpret_cast<const unsigned char*>(GetData()), paramLen);
+                            reinterpret_cast<const unsigned char*>(GetData()), static_cast<int32_t>(paramLen));
 
                         sscanf(str.c_str(), "%d-%d-%d %d:%d:%d", &tmTime.tm_year, &tmTime.tm_mon,
                             &tmTime.tm_mday, &tmTime.tm_hour, &tmTime.tm_min, &tmTime.tm_sec);
